test(seo): move docs-chrome cases into their own file
generate-llms-txt.test.ts is already past the size limit, so growing it
trips the quality gate. Split the new cases out, matching the existing
generate-llms-txt-sitemap.test.ts precedent.
